% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_edgegroupLik.R
\name{addComputeLik}
\alias{addComputeLik}
\title{Adds two likelihood computations}
\usage{
addComputeLik(cl1, cl2)
}
\arguments{
\item{cl1}{[list] :: Output of \code{\link{computeLik}}}

\item{cl2}{[list] :: Output of \code{\link{computeLik}}}
}
\value{
[list] Output with log-likelihood computations added
}
\description{
Adds two likelihood computations
}

