% Generated by roxygen2 (4.0.1): do not edit by hand
\name{aggstat_corr}
\alias{aggstat_corr}
\title{Compute sufficient statistics for correlated-model estimation}
\usage{
aggstat_corr(NetM, adja1, adja2)
}
\arguments{
\item{NetM}{[\code{\link{NetworkModel}}] :: Model with respect to which structure would be computed}

\item{adja1}{[array-int] :: Adjacency matrix or arrays 1}

\item{adja2}{[array-int] :: Adjacency matrix or arrays 2}
}
\value{
[list] :: List with following entries
\itemize{
 \item n -- [vector-int] :: Edge group sizes
 \item names -- [vector-char] :: Dyad group IDs
 \item C - [matrix-int] :: columns are counts of 11, 10, 01, 00 cases respectively
}
}
\description{
Compute sufficient statistics for correlated-model estimation
}

