% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_edgegroupLik.R
\name{aggstat_dendiff}
\alias{aggstat_dendiff}
\title{Compute sufficient statistics for density-difference estimation}
\usage{
aggstat_dendiff(NetM, adja1, adja2)
}
\arguments{
\item{NetM}{[\code{\link{NetworkModel}}] :: Model with respect to which structure would be computed}

\item{adja1}{[array-int] :: Adjacency matrix or arrays 1}

\item{adja2}{[array-int] :: Adjacency matrix or arrays 2}
}
\value{
[list] :: List with following entries
\itemize{
 \item n -- [vector-int] :: Edge group sizes
 \item x -- [vector-int] :: Dyad group edge presence counts in input 1
 \item y -- [vector-int] :: Dyad group edge presence counts in input 2
 \item names -- [vector-char] :: Dyad group IDs
}
}
\description{
Compute sufficient statistics for density-difference estimation
}

