% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_edgegroupLik.R
\name{aggstat_single}
\alias{aggstat_single}
\title{<What does this function do>}
\usage{
aggstat_single(NetM, adja)
}
\arguments{
\item{NetM}{temp}

\item{adja}{temp}
}
\value{
temp
}
\description{
<What does this function do>
}

