% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_hrg_fx.R
\name{anc_table_from_expanded_children}
\alias{anc_table_from_expanded_children}
\title{Compute ancestor table}
\usage{
anc_table_from_expanded_children(expanded.children, nn)
}
\arguments{
\item{expanded.children}{List of expanded children}

\item{nn}{Number nodes in tree}
}
\value{
Ancestor table (matrix)
}
\description{
Compute ancestor table
}

