% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_hrg_fx.R
\name{closest_ancestor}
\alias{closest_ancestor}
\title{Compute ancestor_table from a tree}
\usage{
closest_ancestor(tree)
}
\arguments{
\item{tree}{[list] :: CMN tree model object}
}
\value{
[list] :: expanded_children & ancestor_table
}
\description{
Compute ancestor_table from a tree
}

