% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_hrg_fx.R
\name{closest_ancestor_new}
\alias{closest_ancestor_new}
\title{Compute ancestor_table from a tree}
\usage{
closest_ancestor_new(tree)
}
\arguments{
\item{tree}{[\code{\link{NetworkModelHRG}}] :: Input tree}
}
\value{
[list] :: expanded_children & ancestor_table
}
\description{
Compute ancestor_table from a tree
}

