% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareNetworks.R
\name{compareNetworks}
\alias{compareNetworks}
\title{Compares two networks}
\usage{
compareNetworks()
}
\value{
p-value of some sort? FIX THIS
}
\description{
This function compares a pair of networks [add stuff on multiple pairs of networks if implemented properly]
}

