% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_compute_lrt.R
\name{compute_cellwise_loglik}
\alias{compute_cellwise_loglik}
\title{Compute the loglikelihood from probabilities and counts}
\usage{
compute_cellwise_loglik(x, n, p)
}
\arguments{
\item{x}{[matrix/array-int] :: observed counts in each cell}

\item{n}{[int] :: Number observations}

\item{p}{[vector-double] :: Corresponding estimated cell probabilities (is just x / n)}
}
\value{
[vector-double] :: Vectorized version of log-likelihood (per edge group)
}
\description{
Compute the loglikelihood from probabilities and counts
}

