% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_compute_lrt.R
\name{compute_loglik_fromPC}
\alias{compute_loglik_fromPC}
\title{Compute the loglikelihood from probabilities and counts}
\usage{
compute_loglik_fromPC(x, n, p)
}
\arguments{
\item{x}{[vector-int] :: observed counts in each edge group}

\item{n}{[vector-int] :: Corresponding size of edge groups (cells)}

\item{p}{[vector-double] :: Corresponding estimated cell probabilities (is just x / n)}
}
\value{
[vector-double] :: Vectorized version of log-likelihood (per edge group)
}
\description{
Compute the loglikelihood from probabilities and counts
}

