% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compute_small_samp_dfadj}
\alias{compute_small_samp_dfadj}
\title{Computes Expected Value of test statistic under small-sample regime}
\usage{
compute_small_samp_dfadj(n, p, mode = "exact")
}
\arguments{
\item{n}{[int] :: Sample size}

\item{p}{[numeric] :: Edge probability}

\item{mode}{temp}
}
\value{
[numeric] :: Returns Expected Value of chi-square test statistic
}
\description{
Computes the expected likelihood ratio test statistic for dyad group of size 'n' and dyad probability 'p'. Asymptotically, this goes to 1, but for small samples, it's not necessarily 1. It is needed since the shape of the null distribution very well approximates a chi-square but with a larger than expected 'df' simply due to being a smaller sample
}

