% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_functions.R
\name{computeEmpDfAdj}
\alias{computeEmpDfAdj}
\title{Estimates the degrees of freedom empirically}
\usage{
computeEmpDfAdj(adjm1, adjm2, NetS, model_type = "default")
}
\arguments{
\item{adjm1}{[matrix/array-int] :: Input adjacency matrix}

\item{adjm2}{[matrix/array-int] :: Input adjacency matrix}

\item{NetS}{[\code{\link{NetworkStruct}}] :: Structure used for testing}

\item{model_type}{[char] :: Model type to compute degrees of freedom for. See ## FIX THIS## for description of model_type.}
}
\value{
[df-vector-double] :: Two columns, 'coradj' and 'ssadj'. 'coradj' is the degrees of freedom for the dyad group provided that
}
\description{
Estimates the degrees of freedom empirically
}

