% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic_computeLik.R
\name{computeLik}
\alias{computeLik}
\title{Compute Likelihood of Network given a model}
\usage{
computeLik(NetM, adja, loglik = TRUE, by_node = FALSE, na.rm = TRUE)
}
\arguments{
\item{NetM}{Network Model}

\item{adja}{Adjacency matrix/array}

\item{loglik}{if true -- give log-likelihood instead of likelihood}

\item{by_node}{temp}

\item{na.rm}{if true -- ignores NAs in the adjacency matrix}
}
\value{
likelihood (or log-likelihood)
}
\description{
Compute Likelihood of Network given a model
}

