% Generated by roxygen2 (4.0.1): do not edit by hand
\name{computePval}
\alias{computePval}
\title{Computes single p-value for likelihood ratio test}
\usage{
computePval(NetS, adja1, adja2, Nobs = 1, pl, output_mode = "pval",
  model_type = "default-slow", verbose = TRUE, vbset = c(1, 0, 0))
}
\arguments{
\item{NetS}{[\code{\link{NetworkStruct}}] :: Model to compute p-value with}

\item{adja1}{[matrix/array] :: Adjacency matrix/array}

\item{adja2}{[matrix/array] :: Adjacency matrix/array}

\item{Nobs}{[int] :: Number of network observations per class (default = 1)}

\item{pl}{[list] :: Simulation/Testing parameters, set by set_sim_param}

\item{output_mode}{[char] :: How to output results? Options are:
\itemize{
\item 'chisq' :: Only gives the asymptotically chi-square test statistic
\item 'nodal' :: Outputs break down of test statistic into each vertex
\item 'pval' :: Outputs p-value based on chi-square distribution
}}

\item{model_type}{[char] :: What model type to fit? Options are:
\itemize{
\item 'default' :: Uses the standard hypothesis test
\item 'default-slow' :: Uses the standard hypothesis test, but slower (but cleaner) R code
\item 'densitydiff' :: Uses version where null allows for a global additive parameter in probability difference
\item 'correlated' :: Uses version where null allows for correlated models
}}

\item{verbose}{[logical] :: Prints progress if TRUE}

\item{vbset}{[vector-int] :: See \link{OutputSettings}
Regular-verbose -- outputs general information for processing the NetworkStructList case
High-verbose -- outputs dots for each struct processed}
}
\value{
[] :: A matrix (or a list of matrices) of p-values (depending on the testing parameters)
}
\description{
This function takes the input network structure (or a list of network structures), and computes test statistics for the likelihood ratio test for the two input adjacency matrices (or arrays).
}
\details{
This function computes a single p-value per network structure (it doesn't do any p-value aggregation). Thus, the only entries in 'pl' it considers are 'cc_adj' and 'thres_ignore'.
}

