% Generated by roxygen2 (4.0.1): do not edit by hand
\name{computeTrueDfAdj}
\alias{computeTrueDfAdj}
\title{Compute Theoretical df Adjustment}
\usage{
computeTrueDfAdj(NetM, NetS, hidden_edges = NULL)
}
\arguments{
\item{NetM}{[\code{\link{NetworkModel}}] :: Underlying null model}

\item{NetS}{[\code{\link{NetworkStruct}}] :: Testing model structure}

\item{hidden_edges}{[matrix-logical] :: Entries of matrix are 'true' if they are hidden in the inference process. (so default would be a matrix of falses)}
}
\value{
[vector-double] :: Theoretical df adjustments for each edge group
}
\description{
For a fixed null hypothesis model, and for a fixed network structure, compute the true theoretical df adjustment
}

