% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_hrg_fx.R
\name{depth_from_parents}
\alias{depth_from_parents}
\title{Compute depths of each node}
\usage{
depth_from_parents(parents)
}
\arguments{
\item{parents}{[vector-int] :: Valued from 0 to length; these indicate which index the parent node is at}
}
\value{
[vector-int] :: Depth of each node
}
\description{
Compute depths of each node
}

