% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_SBM.R
\name{EM_SBM_mf}
\alias{EM_SBM_mf}
\title{Fit SBM using mean field approx}
\usage{
EM_SBM_mf(adjm, Nobs, nodeps, edgeps, H, PHI, Niter, stop_thres, verbose)
}
\arguments{
\item{adjm}{temp}

\item{Nobs}{temp}

\item{nodeps}{temp}

\item{edgeps}{temp}

\item{H}{temp}

\item{PHI}{temp}

\item{Niter}{temp}

\item{stop_thres}{temp}

\item{verbose}{temp}
}
\value{
temp
}
\description{
Fit SBM using mean field approx
}

