% Generated by roxygen2 (4.0.1): do not edit by hand
\name{EM_SBM_mf_chooser}
\alias{EM_SBM_mf_chooser}
\title{Fit SBM using mean field approx}
\usage{
EM_SBM_mf_chooser(adjm, Nobs, nodeps, edgeps, H, PHI, Niter, stop_thres, mode,
  verbose)
}
\arguments{
\item{adjm}{[matrix-int] :: Input adjacency matrix}

\item{Nobs}{[int] :: Number of observations per dyad}

\item{nodeps}{[vec-double] :: Initial node probabilities}

\item{edgeps}{[matrix-double] :: Initial between-block probabilities}

\item{H}{[matrix-double] :: Initial value of H}

\item{PHI}{[matrix-double] :: Initial value of PHI}

\item{Niter}{[int] :: Number of iterations max}

\item{stop_thres}{[double] :: Stop condition}

\item{mode}{temp}

\item{verbose}{[logical] :: Output stuff?}
}
\value{
[list] :: EM algorithm output, with the following elements:
\itemize{
\item nodeps -- [vec-double] :: Estimated class probabilities
\item edgeps -- [matrix-double] :: Estimated between-class edge probabilities
\item classes -- [vec-int] :: Estimated class assignments
\item model -- [\code{\link{NetworkModel}}] :: Estimated Network Model
}
}
\description{
Fit SBM using mean field approx
}

