% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_SBM_EMalgos.R
\name{EM_SBM_mf_twoedgeps_nonode}
\alias{EM_SBM_mf_twoedgeps_nonode}
\title{Fit SBM using mean field approx}
\usage{
EM_SBM_mf_twoedgeps_nonode(adjm1, adjm2, Nobs, nodeps, edgeps1, edgeps2, H, PHI,
  Niter, stop_thres, verbose)
}
\arguments{
\item{adjm1}{temp}

\item{adjm2}{temp}

\item{Nobs}{[int] :: Number of observations per dyad}

\item{nodeps}{[vec-double] :: Initial node probabilities}

\item{edgeps1}{temp}

\item{edgeps2}{temp}

\item{H}{[matrix-double] :: Initial value of H}

\item{PHI}{[matrix-double] :: Initial value of PHI}

\item{Niter}{[int] :: Number of iterations max}

\item{stop_thres}{[double] :: Stop condition}

\item{verbose}{[logical] :: Output stuff?}

\item{edgeps}{[matrix-double] :: Initial between-block probabilities}

\item{adjm}{[matrix-int] :: Input adjacency matrix}
}
\value{
[list] :: EM algorithm output, with the following elements: 
\itemize{
\item nodeps -- [vec-double] :: Estimated class probabilities
\item edgeps -- [matrix-double] :: Estimated between-class edge probabilities
\item classes -- [vec-int] :: Estimated class assignments
\item model -- [\code{\link{NetworkModel}}] :: Estimated Network Model
}
}
\description{
Fit SBM using mean field approx
}

