% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_hrg_fx.R
\name{expanded_children_from_tree}
\alias{expanded_children_from_tree}
\title{Extract expanded children for a tree}
\usage{
expanded_children_from_tree(tree)
}
\arguments{
\item{tree}{[\code{\link{NetworkModelHRG}}] :: Input network model}
}
\value{
[list] :: A tree in list format (left child index, right child index)
}
\description{
Extract expanded children for a tree
}

