% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic_extractModel.R
\name{extractModel}
\alias{extractModel}
\title{Generates a random network model from a fixed structure}
\usage{
extractModel(NetS, model_param = set_model_param())
}
\arguments{
\item{NetS}{temp}

\item{model_param}{temp}
}
\value{
temp
}
\description{
Generates a random network model from a fixed structure
}

