% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic_extractStruct.R
\name{extractStruct}
\alias{extractStruct}
\title{Extracts the Edge Partition from a Network Model}
\usage{
extractStruct(NetM)
}
\arguments{
\item{NetM}{NetworkModel object}
}
\value{
NetworkStruct object that corresponds with the input model
}
\description{
Extracts the Edge Partition from a Network Model
}

