% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_SBM.R
\name{fit_SBM}
\alias{fit_SBM}
\title{Fit SBM given adjacency matrix}
\usage{
fit_SBM(adjm, Nobs = 1, control_list = set_fit_param())
}
\arguments{
\item{adjm}{[matrix-int] :: Input adjacency matrix}

\item{Nobs}{[int] :: Number of observations}

\item{control_list}{temp}
}
\value{
[\code{\link{NetworkModel}}] :: Best found fitted model
}
\description{
Fit SBM given adjacency matrix
}

