% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fit_SBM_pair_searchboth}
\alias{fit_SBM_pair_searchboth}
\title{<What does this function do>}
\usage{
fit_SBM_pair_searchboth(adjm1, adjm2, Nobs = 1,
  control_list = set_fit_param())
}
\arguments{
\item{adjm1}{temp}

\item{adjm2}{temp}

\item{Nobs}{temp}

\item{control_list}{temp}
}
\value{
temp
}
\description{
<What does this function do>
}

