% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_fitModel.R
\name{fitModel}
\alias{fitModel}
\title{Estimates the best-fit model for given structure}
\usage{
fitModel(NetS, adja, mode = "default", optim_tries = 10)
}
\arguments{
\item{NetS}{[\code{\link{NetworkStruct}}] :: Edge partition to fit model on}

\item{adja}{[matrix/array] :: Adjacency array or matrix}

\item{mode}{[char] :: Method to fit model -- choices are: "default", "densitydiff", "corr-global-null", "corr-global"}

\item{optim_tries}{[int] :: Number of attempts at optimization of likelihood function}
}
\value{
[\code{\link{NetworkModel}}] :: Returns the best-fit model for each structure. If input is a \code{\link{NetworkStructList}}, output will actually be a list of \code{\link{NetworkModel}}s.
}
\description{
Note: if adja is NULL, will fill in random probabilities drawn from an uniform distribution
}

