% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_dyadgroup_prob}
\alias{get_dyadgroup_prob}
\title{Get dyad group probabilities}
\usage{
get_dyadgroup_prob(NetM)
}
\arguments{
\item{NetM}{[\code{\link{NetworkModel}}] :: Model to get group probabilities of}
}
\value{
[list] :: Two elements:
\itemize{
\item probs -- [vector-double] :: Probabilities for each dyad group
\item names -- [vector-char] :: Names for each dyad group for ID purposes
}
}
\description{
Get dyad group probabilities
}

