% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic_getEdgeProbMat.R
\name{getEdgeProbMat}
\alias{getEdgeProbMat}
\title{Extracts the edge probability matrix for a network model}
\usage{
getEdgeProbMat(NetM, mode = "prob")
}
\arguments{
\item{NetM}{Network model object}

\item{mode}{Either "prob" or "group": if "prob", this function returns an edge probability matrix. Else, it returns a matrix where each cell gets a group id number}
}
\value{
Matrix: Edge Probability Matrix given by the model
}
\description{
Extracts the edge probability matrix for a network model
}

