% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic_getNetType.R
\name{getNetType}
\alias{getNetType}
\title{Extract the type of network model}
\usage{
getNetType(Net)
}
\arguments{
\item{Net}{Object of class NetworkModel (or inherits this type)}
}
\value{
Character -- type of network model
}
\description{
Allowable types are currently: "block", "tree", "latent", "random"
}

