% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_getNnodes.R
\name{getNnodes}
\alias{getNnodes}
\title{Extracts the number of nodes in a network model}
\usage{
getNnodes(Net)
}
\arguments{
\item{Net}{[\code{\link{NetworkModel}} OR \code{\link{NetworkStruct}}] :: Network object}
}
\value{
[int] :: Number nodes in network
}
\description{
Extracts the number of nodes in a network model
}

