% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_network.R
\name{graph_laplacian}
\alias{graph_laplacian}
\title{<What does this function do>}
\usage{
graph_laplacian(adjm)
}
\arguments{
\item{adjm}{temp}
}
\value{
temp
}
\description{
<What does this function do>
}

