% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_edgegroupLik.R
\name{hCorr_paramToProb}
\alias{hCorr_paramToProb}
\title{<What does this function do>}
\usage{
hCorr_paramToProb(rho, a, b = a)
}
\arguments{
\item{rho}{temp}

\item{a}{temp}

\item{b}{temp}
}
\value{
temp
}
\description{
<What does this function do>
}

