% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_edgegroupLik.R
\name{hCorr_probToParam}
\alias{hCorr_probToParam}
\title{<What does this function do>}
\usage{
hCorr_probToParam(pt)
}
\arguments{
\item{pt}{temp}
}
\value{
temp
}
\description{
<What does this function do>
}

