% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_edgegroupLik.R
\name{hide_edges}
\alias{hide_edges}
\title{Hides a random set of edges}
\usage{
hide_edges(adjm, frac = 0.1)
}
\arguments{
\item{adjm}{Input adjacency matrix}

\item{frac}{Fraction of edges to hide}
}
\value{
Adjacency matrix with hidden edges
}
\description{
Hides a random set of edges
}

