% Generated by roxygen2 (4.0.1): do not edit by hand
\name{HRG_depthFromParents}
\alias{HRG_depthFromParents}
\title{Compute depths of each node}
\usage{
HRG_depthFromParents(parents)
}
\arguments{
\item{parents}{[vector-int] :: Index location of parent node. Value of 0 indicates that this is the root node.}
}
\value{
[vector-int] :: Depth of each node (root node has depth 1)
}
\description{
Compute depths of each node
}

