% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_hrg_fx.R
\name{HRG_expandedChildren}
\alias{HRG_expandedChildren}
\title{Extract expanded children for a tree}
\usage{
HRG_expandedChildren(tree)
}
\arguments{
\item{tree}{[\code{\link{NetworkModelHRG}}] :: Input network model}
}
\value{
[list] :: A tree in list format (left child index, right child index)
}
\description{
Extract expanded children for a tree
}

