% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_edgegroupLik.R
\name{llFx_cnull}
\alias{llFx_cnull}
\title{Computes Log-likelihood for null-hypothesis correlation case}
\usage{
llFx_cnull(t, C, n)
}
\arguments{
\item{t}{[vector-double] :: c(rho, a_v)}

\item{C}{[matrix-int] :: Dyad group counts, in format: cbind(c11, c10, c01, c00)}

\item{n}{[vector-int] :: Dyad group sizes}
}
\value{
[double] :: Log-likelihood at current point
}
\description{
Computes Log-likelihood for null-hypothesis correlation case
}

