% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_edgegroupLik.R
\name{llFx_dendiff}
\alias{llFx_dendiff}
\title{Compute log-likelihood function for density-difference}
\usage{
llFx_dendiff(t, x, y, n)
}
\arguments{
\item{t}{[] :: c(b,a) - b is single value; a is vector}

\item{x}{[] :: edge counts in net 1}

\item{y}{[] :: edge counts in net 2}

\item{n}{[] :: total edge group size}
}
\value{
[] :: value of log-likelihood
}
\description{
Compute log-likelihood function for density-difference
}

