% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_edgegroupLik.R
\name{llGrFx_dendiff}
\alias{llGrFx_dendiff}
\title{Compute gradient vector of log-likelihood in density-difference model}
\usage{
llGrFx_dendiff(t, x, y, n)
}
\arguments{
\item{t}{[] :: c(b,a) - b is single value; a is vector}

\item{x}{[] :: edge counts in net 1}

\item{y}{[] :: edge counts in net 2}

\item{n}{[] :: total edge group size}
}
\value{
[] :: gradient vector
}
\description{
Compute gradient vector of log-likelihood in density-difference model
}

