% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lower_diag}
\alias{lower_diag}
\title{Obtains the lower diagonal area of a square matrix}
\usage{
lower_diag(nn)
}
\arguments{
\item{nn}{[int] :: Number of rows in square matrix}
}
\value{
[vector-int] :: Indices for lower-diagonal area
}
\description{
Obtains the lower diagonal area of a square matrix
}

