% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mult_highcrit}
\alias{mult_highcrit}
\title{Compute higher-critiism test statistic}
\usage{
mult_highcrit(x, alpha_min = 10^(-8), alpha_max = 0.2)
}
\arguments{
\item{x}{[vector-double] :: P-values}

\item{alpha_min}{[double] :: Minimal alpha to maximize over}

\item{alpha_max}{[double] :: Maximal alpha to maximize over}
}
\value{
[double] :: Test statistic
}
\description{
Compute higher-critiism test statistic
}

