% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiple_testing_fx.R
\name{mult_pearson}
\alias{mult_pearson}
\title{Computes pearson-like test statistic (or p-value)}
\usage{
mult_pearson(x, adjust = NULL)
}
\arguments{
\item{x}{[vector-double] :: P-values}

\item{adjust}{[logical] :: If NULL, function returns test statistic. Otherwise, this takes on a multiplier to variance.}
}
\value{
[double] :: Adjusted test statistic
}
\description{
Computes p-value by assuming normality (using larger variance estimate)
Larger variance estiamte is bounded by 1/n + cor, but uses 1/n * adjust as variance.
}

