% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing_fx.R
\name{mult_quantileq}
\alias{mult_quantileq}
\title{Compute quantile test statistic (Q)}
\usage{
mult_quantileq(x, alpha = 0.05)
}
\arguments{
\item{x}{[vector-double] :: Input p-values}

\item{alpha}{[double] :: Alpha parameter in statistic}
}
\value{
[double] :: Test statistic
}
\description{
Compute quantile test statistic (Q)
}

