% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing_fx.R
\name{mult_quantiler}
\alias{mult_quantiler}
\title{Compute quantile test statistic (R)}
\usage{
mult_quantiler(x, alpha_min = 0.05, alpha_max = 1)
}
\arguments{
\item{x}{[vector-double] :: Input p-values}

\item{alpha_min}{[double] :: Lower threshold for alpha parameter}

\item{alpha_max}{[double] :: Upper threshold for alpha parameter}
}
\value{
[double] :: Test statistic
}
\description{
Compute quantile test statistic (R)
}

