% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{NetworkModel}
\alias{NetworkModel}
\title{Instantiates an object of class NetworkModel}
\usage{
NetworkModel(model_params = set_model_param())
}
\arguments{
\item{model_params}{[list; DEFAULT = \code{\link{set_model_param}}()] :: Model parameters}
}
\value{
[NetworkModel] :: A representation of a model generated as specified
}
\description{
Instantiates an object of class NetworkModel
}

