% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{NetworkModelHRG}
\alias{NetworkModelHRG}
\title{Constructor for HRG network model}
\usage{
NetworkModelHRG(model_params = set_model_param())
}
\arguments{
\item{model_params}{[list; DEFAULT = \code{\link{set_model_param}}()] :: Model parameters}
}
\value{
[NetworkModelHRG] :: A representation of the generated model
}
\description{
This will generate a network model with random class assignments and class probabilities (unless otherwise specified)
}

