% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{NetworkModelPair}
\alias{NetworkModelPair}
\title{Instantiates an object of class NetworkModelPair}
\usage{
NetworkModelPair(m1, m2 = NULL, is_null = FALSE, model_type = "default",
  addl_param = list())
}
\arguments{
\item{m1}{[\code{\link{NetworkModel}} OR list] :: This can either be a model object, or a list containing model parameters}

\item{m2}{[\code{\link{NetworkModel}} OR list] :: This can either be a model object, or a list containing model parameters}

\item{is_null}{[logical] :: If TRUE, m2 is ignored (since the null hypothesis means that both models are identical)}

\item{model_type}{[char] :: Can be 'default', 'correlated', or 'densitydiff'}

\item{addl_param}{[list] :: Additional parameters for correlated or densitydiff models}
}
\value{
[NetworkModelPair] :: An object containing information about two network models.
}
\description{
This is done by providing both network models. If m2 is not given and is_null is FALSE, there is an error. Otherwise, m2 will be ignored if is_null is TRUE. (ie null hypothesis means both models are the same, so m1)
There is also no check that m1 and m2 are on the same network size, but they should be. -- add this in...
}

