% Generated by roxygen2 (4.0.1): do not edit by hand
\name{NetworkModelRND}
\alias{NetworkModelRND}
\title{Constructor for RND network model}
\usage{
NetworkModelRND(model_params = set_model_param())
}
\arguments{
\item{model_params}{[list; DEFAULT = \code{\link{set_model_param}}()] :: Model parameters}
}
\value{
[NetworkModelRND] :: A representation of the generated model
}
\description{
This will generate a network model with random class assignments and class probabilities (unless otherwise specified)
}

