% Generated by roxygen2 (4.0.1): do not edit by hand
\name{NetworkModelSBM}
\alias{NetworkModelSBM}
\title{Constructor for SBM network model}
\usage{
NetworkModelSBM(model_params = set_model_param())
}
\arguments{
\item{model_params}{[list; DEFAULT = \code{\link{set_model_param}}()] :: Model parameters}
}
\value{
[NetworkModelSBM] :: A representation of the generated model
}
\description{
This will generate a network model with random class assignments and class probabilities (unless otherwise specified)
}

