% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{NetworkStruct}
\alias{NetworkStruct}
\title{Instantiates an object of class NetworkStruct}
\usage{
NetworkStruct(model_params = set_model_param())
}
\arguments{
\item{model_params}{[list; DEFAULT = \code{\link{set_model_param}}()] :: Model parameters}
}
\value{
[NetworkStruct] :: A representation of the generated model structure
}
\description{
Instantiates an object of class NetworkStruct
}

