% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{NetworkStructList}
\alias{NetworkStructList}
\title{Constructor for class NetworkStructList -- a list of network model structure information.}
\usage{
NetworkStructList(Nmodels = 10, model_params = set_model_param(),
  NetMPair = NULL)
}
\arguments{
\item{Nmodels}{[int] :: Number of random edge partitions to generate}

\item{model_params}{[list; DEFAULT = \code{\link{set_model_param}}()] :: Model parameters}

\item{NetMPair}{[\code{\link{NetworkModelPair}}] :: If not NULL, these models are used, and its structures extracted}
}
\value{
[NetworkStructList] :: A representation of the generated model structure
}
\description{
These are the random edge partitions used in the network hypothesis testing.
}

