% Generated by roxygen2 (4.0.1): do not edit by hand
\name{NetworkStructRND}
\alias{NetworkStructRND}
\title{Constructor for RND network structure}
\usage{
NetworkStructRND(model_params = set_model_param(), NetM = NULL)
}
\arguments{
\item{model_params}{[list; DEFAULT = \code{\link{set_model_param}}()] :: Model parameters}

\item{NetM}{[\code{\link{NetworkModelRND}}] :: If not NULL, this model is used, and its structure extracted}
}
\value{
[NetworkStructRND] :: A representation of the generated model structure
}
\description{
This will generate a network model with random class assignments and class probabilities (unless otherwise specified)
}

