% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_heatmap.R
\name{plot_edgegroup}
\alias{plot_edgegroup}
\title{Plot the Edge Group structure of a network model}
\usage{
plot_edgegroup(NetM, cols = 1:100, ...)
}
\arguments{
\item{NetM}{[\code{\link{NetworkModel}}] :: Model to plot edge group of}

\item{cols}{[vector-int OR vector-char] :: Specify coloring scheme for edge groups}

\item{...}{[] :: Further parameters to pass on to the plot function}
}
\value{
[int] :: Ignore output; a plot is produced
}
\description{
Plot the Edge Group structure of a network model
}

