% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_heatmap.R
\name{plot_probmatrix}
\alias{plot_probmatrix}
\title{Plots a heatmap of a probability matrix}
\usage{
plot_probmatrix(pm, add_legend = TRUE, is_dm = FALSE, only_legend = FALSE,
  colors = rainbow(n = 200, s = 1, v = 1, start = 0, end = 1/3, alpha = 1),
  ...)
}
\arguments{
\item{pm}{[matrix-double] :: Probability matrix}

\item{add_legend}{[logical; DEFAULT = T] :: If TRUE, adds a legend plot}

\item{is_dm}{[logical; DEFAULT = F] :: If TRUE, input matrix is on a [-1,1] scale (as a difference of probability matrices)}

\item{only_legend}{[logical; DEFAULT = F] :: If TRUE: ONLY plot the legend}

\item{colors}{[vector-char] :: Colors to use in the heat map (as a vector of hex values)}

\item{...}{[] :: Other variables to be passed to 'plot'}
}
\value{
[nothing] :: Produces a plot (standard R plot functions)
}
\description{
Generates a heatmap of a probability matrix. This allows the probabilities to also be a difference between two probability matrices, and allows custom input of color scheme.
}

