% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_sampleNetwork.R
\name{sampleNetwork}
\alias{sampleNetwork}
\title{Sample a network from a network model}
\usage{
sampleNetwork(NetM, Nobs = 1, Nsim = 1)
}
\arguments{
\item{NetM}{[\code{\link{NetworkModel}}] :: Model to sample network from}

\item{Nobs}{[int; DEFAULT = 1] :: Number of network observations to simulate}

\item{Nsim}{[int; DEFAULT = 1] :: Number of simulations}
}
\value{
[array OR list-array OR list-(array OR list-array)] :: List or array of adjacency matrices. If calling on a \code{\link{NetworkModelPair}}, the output is a list of output from calling this function on each of the two specific network models.
}
\description{
Nobs is the number of network observations to simulate per simulation. This is built in here since callling sampleNetwork many times is slow (since it requires re-computing the edge probability matrix, which is the same each time given the same network model). There are several cases: 
If Nsim = 1 -> Result is an array (with third dimension equal to Nobs)
If Nsim > 1 -> Result is a list of arrays
}

