% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_edgegroupLik.R
\name{set_dyadgroup_prob}
\alias{set_dyadgroup_prob}
\title{Set dyad group probabilities}
\usage{
set_dyadgroup_prob(NetM, ids, probs)
}
\arguments{
\item{NetM}{[\code{\link{NetworkModel}}] :: Model to reassign probabilities of}

\item{ids}{[vector-char OR vector-int] :: Dyad group IDs}

\item{probs}{[vector-double] :: Probabilities to assign to dyad groups}
}
\value{
[\code{\link{NetworkModel}}] :: Adjusted network model with reassigned probabilities
}
\description{
Set dyad group probabilities
}

