% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parameter_lists.R
\name{set_fit_param}
\alias{set_fit_param}
\title{Set model fitting parameters}
\usage{
set_fit_param(Ntries = 10, verbose = 0, SBM_start = "spectral",
  SBM_method = "mf", SBM_Nclass = 3, SBM_EM_Niter = 100,
  SBM_EM_stopthres = 1e-05)
}
\arguments{
\item{Ntries}{temp}

\item{verbose}{temp}

\item{SBM_start}{temp}

\item{SBM_method}{temp}

\item{SBM_Nclass}{temp}

\item{SBM_EM_Niter}{temp}

\item{SBM_EM_stopthres}{temp}
}
\value{
[list] :: A list of parameters
}
\description{
This function sets up the (list of) parameters to be passed into model fitting functions. There are default parameters that are used if this function is called with no arguments.
}

