% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parameter_lists.R
\name{set_model_param}
\alias{set_model_param}
\title{Set network model parameters}
\usage{
set_model_param(Nnodes = 30, type = "block", pairtype = NULL,
  pmin = 0.03, pmax = 0.97, block_nclass = 3, block_avgdensity = NULL,
  block_assign = NULL, block_probs = NULL, random_ngroups = 10,
  tree_type = "random", latent_dim = 3, latent_nclass = 3,
  latent_sdcenter = 5, latent_isgennorm = TRUE, dd_param_add = 0.2,
  c_param_corr = 0.2)
}
\arguments{
\item{Nnodes}{[int] :: Number of nodes in the network model}

\item{type}{[char; ALLOWED = c("none", "block", "tree", "latent", "random")] :: type of network model
\itemize{
 \item none -- Probably not useful; can be used to construct relatively empty model class for whatever purpose...
 \item block -- Stochastic Block Model
 \item tree -- Hierarchical Random Graph
 \item latent -- Latent Space Model. Note that this type of model CANNOT generate random edge partitions since edge groups will be size 1 each (almost surely)
 \item random -- A model where each edge belongs to one of a number of classes, and the edge probability depends solely on this assignment. Should really only be used to generate random edge partitions.
}}

\item{pairtype}{[char] :: This corresponds to model_type. If NULL, this is not a model pair. If non-null, it should be one of:
\itemize{
\item 'densitydiff' -- Model with logit-odds additive difference in density
\item 'correlated-null' -- Model with same marginal distribution but correlated
\item 'correlated' -- Model with possibly different marginals but correlated
}}

\item{pmin}{[double] :: Minimal possible edge probability}

\item{pmax}{[double] :: Maximal possible edge probability}

\item{block_nclass}{[int] :: Number of blocks in block model}

\item{block_avgdensity}{[double] :: Set average density in block model (ignored if NULL)}

\item{block_assign}{[vector-int] :: block assignments}

\item{block_probs}{[matrix-double] :: block model probabilities}

\item{random_ngroups}{[int] :: Number of groups for completely random edge partition}

\item{tree_type}{[char; ALLOWED = c("left", "random", "original")] :: Randomization on structure of tree
\itemize{
 \item left -- This is a left-branching tree (all splits are on the left branch), but the nodes are in a random order
 \item random -- This is a tree where a random terminal node is chosen for a split. Tends to randomize into more balanced trees
 \item original -- This is the most balanced possible tree, with node order shuffled.
}}

\item{latent_dim}{[int] :: Dimension of latent space in latent space models}

\item{latent_nclass}{[int] :: Number of clusters in latent space model}

\item{latent_sdcenter}{[double] :: SD on centers of latent space model}

\item{latent_isgennorm}{[logical] :: If TRUE: Uses normal distribution for latent locations. Otherwise, uses uniform distribution.}

\item{dd_param_add}{[double] :: Additive value of parameter in density-difference model}

\item{c_param_corr}{[double] :: Correlation parameter in correlated model}
}
\value{
[list] :: A list of parameters
}
\description{
This function sets up the (list of) model parameters to be passed into model generation functions (eg. sets the max number of blocks in a SBM). There are default parameters that are used if this function is called with no arguments.
}

