% Generated by roxygen2 (4.0.1): do not edit by hand
\name{set_sim_param}
\alias{set_sim_param}
\title{Set up simulation parameters}
\usage{
set_sim_param(cc_adj = c(0, 2), thres_ignore = c(5, 10), alphas = 0.05,
  n_structs = c(1, 25, 50, 100), pval_fx_names = c("mult_bonferroni",
  "mult_highcrit", "mult_fisher"), pval_sim_null = c(FALSE, TRUE, TRUE),
  fitstruct_method = "random")
}
\arguments{
\item{cc_adj}{[vector-double] :: Amount of SE's away from the correlation estimate used (how conservative? 0 means no adjustment (and requires large sample for guarantees; larger values give a conservative p-value))}

\item{thres_ignore}{[vector-int] :: Ignore edge groups with fewer than this many edges}

\item{alphas}{[vector-double] :: Size(s) of the hypothesis test (probability of reject given true null)}

\item{n_structs}{[vector-int] :: Number(s) of edge partitions to use for testing}

\item{pval_fx_names}{[vector-char] :: Names of functions that do p-value adjustment}

\item{pval_sim_null}{[vector-logical] :: Is simulation of the null distribution necessary?}

\item{fitstruct_method}{[char] :: Determines how to use/reuse the fitting structures:
\itemize{
\item 'random' -- Take a random subset at different sizes
\item 'recycle' -- Reuse fixed structures in a fixed manner
\item 'single' -- Do not reuse structures; generate much more
}}
}
\value{
[list] :: A list of parameters, structured as follows
\itemize{
\item cc_adj -- [vector-double] :: Amount of SE's away from the correlation estimate used (how conservative? 0 means no adjustment (and requires large sample for guarantees; larger values give a conservative p-value))
\item thres_ignore -- [vector-int] :: Ignore edge groups with fewer than this many edges
\item alphas -- [vector-double] :: Size(s) of the hypothesis test (probability of reject given true null)
\item n_structs -- [vector-int] :: Number(s) of edge partitions to use for testing
\item pval_adj -- [list] :: Information about p-value adjustment functions
  \itemize{
   \item fx -- [named_list-functions] :: This is a named list of functions
   \item simnull -- [vector-logical] :: This is a corresponding vector -- TRUE if the null distribution needs to be simulated.
  }
\item struct_needed -- [int] :: Total number of random structures needed
\item struct_indices -- [list-vector-int] :: Indices for the structure list for each iteration (each value in n_structs)
}
}
\description{
This function creates a list of simulation parameters that are desired.
}

