% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setup_array}
\alias{setup_array}
\title{Create a named array matching a list}
\usage{
setup_array(pl)
}
\arguments{
\item{pl}{[list] :: List of simulation parameters, as given by \code{\link{set_sim_param}}()}
}
\value{
[array] :: NA-filled named array
}
\description{
Create a named array matching a list
}

