% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_hyptest.R
\name{sim_critvals}
\alias{sim_critvals}
\title{Simulates critical values for each type of multiple testing adjustment}
\usage{
sim_critvals(NetMPair, Nsim = 500, Nobs = 1, fit_NetSList = NULL,
  fit_models_params = set_model_param(), pl = set_sim_param(),
  verbose = 0, vbset = c(1, 0, 0))
}
\arguments{
\item{NetMPair}{[\code{\link{NetworkModelPair}}] :: Pair of network models: These serve as the 'null hypothesis' models from which to generate critical values from}

\item{Nsim}{[int; DEFAULT = 500] :: Number of simulations}

\item{Nobs}{[int; DEFAULT = 1] :: Number of network observations}

\item{fit_NetSList}{[\code{\link{NetworkStructList}}] :: A list of fitting structures. If NULL, an appropriate number of structures will be generated using input of fit_models_params}

\item{fit_models_params}{[list; DEFAULT = \code{\link{set_model_param}}()] :: parameter settings for fitted structures; this can be input instead of a list of pre-generated structures}

\item{pl}{[list] :: List of simulation parameters, typically generated by a call to set_sim_param()}

\item{verbose}{[int; DEFAULT = 0] :: Verbosity level, output stuff if > 2.}

\item{vbset}{[vector-int] :: Passes to \code{\link{sim_hyptest}}; only iterates output headers -> passes exact same verbosity settings to \code{\link{sim_hyptest}}.}
}
\value{
[list] :: list of arrays of critical values
}
\description{
Simulates critical values for each type of multiple testing adjustment
}

