% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim_hyptest}
\alias{sim_hyptest}
\title{Do simulations for the hypothesis testing, using the new code.}
\usage{
sim_hyptest(gen_NetMPair, fit_NetSList = NULL,
  fitm_params = set_model_param(), adjm_list = NULL, Nobs = 1,
  Nsim = 100, pl = set_sim_param(), verbose = TRUE, vbset = c(1, 0, 0))
}
\arguments{
\item{gen_NetMPair}{[\code{\link{NetworkModelPair}}] :: Generating pair of models}

\item{fit_NetSList}{[\code{\link{NetworkStructList}}] :: List of fixed fitting structures (can be NULL -> randomly generate given parameters)}

\item{fitm_params}{[list] :: Output of \code{\link{set_model_params}}, specify the type of fitting model}

\item{adjm_list}{[list-list-array] :: Can input a list of adjacency matrices. Format is a list of list of pairs of arrays. The generating model pair will be ignored if this is provided.}

\item{Nobs}{[int; DEFAULT = 1] :: Number network observations (default = 1)}

\item{Nsim}{[int; DEFAULT = 100] :: Number of simulations to do}

\item{pl}{[list; DEFAULT = \link{set_sim_param}()] :: Parameter list for testing procedure}

\item{verbose}{[logical] :: Output progress?}

\item{vbset}{[vector-int] :: passes along to \code{\link{computePval}}}
}
\value{
[list] :: List of results ## TODO: Format of results?
}
\description{
Do simulations for the hypothesis testing, using the new code.
}

