% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim_power_rpart}
\alias{sim_power_rpart}
\title{Simulates the power for different pairs of generating models}
\usage{
sim_power_rpart(GL, NL, FL, attrib, descrip, outfile, Nsim = 500,
  Nsim_crit = 500, Nobs = 1, verbose = 0, pl = set_sim_param(),
  vbset = c(1, 0, 0))
}
\arguments{
\item{GL}{[list] :: List of generating models (NetModPairs)}

\item{NL}{[list] :: List of null-hypothesis models (NetModPairs)}

\item{FL}{[list] :: List of fitting model parameters}

\item{attrib}{[dataframe] :: Parameters of simulations}

\item{descrip}{[char] :: Description of simulation}

\item{outfile}{[char] :: Filename to store simulation results}

\item{Nsim}{[int; DEFAULT = 500] :: Number of simulations}

\item{Nsim_crit}{[int; DEFAULT = 500] :: Number of simulations for critical values}

\item{Nobs}{[int; DEFAULT = 1] :: Number of network observations}

\item{verbose}{[int; DEFAULT = 0] :: Verbosity level (from 0 to 3), larger means more output}

\item{pl}{[list] :: List of simulation parameters, typically generated by a call to set_sim_param()}

\item{vbset}{temp}
}
\value{
[list] :: List of power-simulation results
}
\description{
Simulates the power for different pairs of generating models
}

