% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_hyptest.R
\name{sim_subsetresults}
\alias{sim_subsetresults}
\title{Extracts results matching a subset of parameters}
\usage{
sim_subsetresults(rl, pl, cc_adj, thres_ignore, alphas, n_structs)
}
\arguments{
\item{rl}{[List] :: Should be output of sim_hyptest (or something similar)}

\item{pl}{[List] :: List of simulation parameters, as given by set_sim_param()}

\item{cc_adj}{[int] :: Amount of SE's away from the correlation estimate used (how conservative? 0 means no adjustment (and requires large sample for guarantees; larger values give a conservative p-value))}

\item{thres_ignore}{[int] :: Ignore edge groups with fewer than this many edges}

\item{alphas}{[int] :: Size of test}

\item{n_structs}{[int] :: Number of edge partitions to use for testing}
}
\value{
[matrix] :: A matrix of results of a certain type
}
\description{
Extracts results matching a subset of parameters
}

