% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_SBM.R
\name{specClust}
\alias{specClust}
\title{Runs spectral clustering on an input matrix}
\usage{
specClust(adjm, evs, Nclass, Ntries, NStartPerTry = 2)
}
\arguments{
\item{adjm}{[matrix-int] :: Input adjacency matrix}

\item{evs}{temp}

\item{Nclass}{temp}

\item{Ntries}{[int] :: Number of attempts}

\item{NStartPerTry}{temp}
}
\value{
[\code{\link{NetworkModel}}] :: Output bestfit model
}
\description{
Runs spectral clustering on an input matrix
}

