% Generated by roxygen2 (4.0.1): do not edit by hand
\name{symmetrize_mat}
\alias{symmetrize_mat}
\title{Symmetrizes matrix by using upper-triangular portion}
\usage{
symmetrize_mat(mat)
}
\arguments{
\item{mat}{[matrix] :: Input matrix}
}
\value{
[matrix] :: Filled matrix (with lower triangular portion replaced by the upper triangular portion).
}
\description{
Symmetrizes matrix by using upper-triangular portion
}

