% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_hrg_fx.R
\name{tree_from_parents}
\alias{tree_from_parents}
\title{Generate tree object from parent vector}
\usage{
tree_from_parents(parents)
}
\arguments{
\item{parents}{[vector-int] :: Index location of parent node. Value of 0 indicates that this is the root node.}
}
\value{
[list] :: A tree in list format (left child index, right child index)
}
\description{
Generate tree object from parent vector
}

