% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_edgegroupLik.R
\name{aggstat_corr}
\alias{aggstat_corr}
\title{<What does this function do>}
\usage{
aggstat_corr(NetM, adja1, adja2)
}
\arguments{
\item{NetM}{temp}

\item{adja1}{temp}

\item{adja2}{temp}
}
\value{
temp
}
\description{
<What does this function do>
}

