% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_edgegroupLik.R
\name{aggstat_dendiff}
\alias{aggstat_dendiff}
\title{Compute sufficient statistics for density-difference estimation}
\usage{
aggstat_dendiff(NetM, adja1, adja2)
}
\arguments{
\item{NetM}{[] :: model}

\item{adja1}{[] :: adjm 1}

\item{adja2}{[] :: adjm 2}
}
\value{
[] :: list of statistics: n, x, y -- n = total edges in edge group, x,y = edge appearance count in each edge group
}
\description{
Compute sufficient statistics for density-difference estimation
}

