% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_edgegroupLik.R
\name{aggstat_single}
\alias{aggstat_single}
\title{Compute sufficient statstics for a single network}
\usage{
aggstat_single(NetM, adja)
}
\arguments{
\item{NetM}{[NetworkModel] :: Model with respect to which structure would be computed}

\item{adja}{[array-int] :: Adjacency matrix or arrays}
}
\value{
[list] :: List with following entries
\itemize{
 \item n -- [vector-int] :: Edge group sizes
 \item x -- [vector-int] :: Dyad group edge presence counts
 \item names -- [vector-char] :: Dyad group IDs
}
}
\description{
Compute sufficient statstics for a single network
}

