% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compareNetworks}
\alias{compareNetworks}
\title{Compares two networks}
\usage{
compareNetworks()
}
\value{
p-value of some sort? FIX THIS
}
\description{
This function compares a pair of networks [add stuff on multiple pairs of networks if implemented properly]
}

