% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/network_functions.R
\name{computeDist}
\alias{computeDist}
\title{Computes a distance between two network models}
\usage{
computeDist(NetM1, NetM2, type = "KLsym")
}
\arguments{
\item{NetM1}{[\code{\link{NetworkModel}}] :: First network model}

\item{NetM2}{[\code{\link{NetworkModel}}] :: Second network model}

\item{type}{[char; ALLOWED = c("KL", "KLsym")] :: What distance measure to use? \cr
\itemize{
  \item KL -- KL distance
  \item KLsym -- symmetric KL distance (average of KL distances in both ways)
}}
}
\value{
[double] :: distance value
}
\description{
Computes a distance between two network models
}

