% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_hrg_fx.R
\name{depth_from_parents}
\alias{depth_from_parents}
\title{Compute depths of each node}
\usage{
depth_from_parents(parents)
}
\arguments{
\item{parents}{[vector-int] :: Index location of parent node. Value of 0 indicates that this is the root node.}
}
\value{
[vector-int] :: Depth of each node (root node has depth 1)
}
\description{
Compute depths of each node
}

