% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractModel}
\alias{extractModel}
\title{Generates a random network model from a fixed structure}
\usage{
extractModel(NetS, model_param = set_model_param())
}
\arguments{
\item{NetS}{[\code{\link{NetworkStruct}}] :: Input network structure}

\item{model_param}{[list] :: Parameters generated from \code{\link{set_model_param}}}
}
\value{
[\code{\link{NetworkModel}}] :: Network Model extracted from input structure
}
\description{
Generates a random network model from a fixed structure
}

