% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_SBM.R
\name{fit_SBM}
\alias{fit_SBM}
\title{<What does this function do>}
\usage{
fit_SBM(adjm, Nobs)
}
\arguments{
\item{adjm}{temp}

\item{Nobs}{temp}
}
\value{
temp
}
\description{
<What does this function do>
}

