% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_SBM.R
\name{fit_SBM}
\alias{fit_SBM}
\title{Fit SBM given adjacency matrix}
\usage{
fit_SBM(adjm, Nobs = 1, Nclass = 3, Niter = 100)
}
\arguments{
\item{adjm}{temp}

\item{Nobs}{temp}

\item{Nclass}{temp}

\item{Niter}{temp}
}
\value{
temp
}
\description{
Fit SBM given adjacency matrix
}

