% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fit_SBM}
\alias{fit_SBM}
\title{Fit a SBM given an input adjacency matrix (pair) and possibly starting model}
\usage{
fit_SBM(adjm, Nobs = 1, control_list = set_fit_param())
}
\arguments{
\item{adjm}{[matrix-int] :: Input adjacency matrix}

\item{Nobs}{[int] :: Number of observations}

\item{control_list}{[list] :: List of control parameters -- see \code{\link{set_fit_param}}}
}
\value{
[\code{\link{NetworkModel}}] :: Best found fitted model
}
\description{
Fit a SBM given an input adjacency matrix (pair) and possibly starting model
}

