% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_edgegroupLik.R
\name{get_dyadgroup_prob}
\alias{get_dyadgroup_prob}
\title{<What does this function do>}
\usage{
get_dyadgroup_prob(NetM)
}
\arguments{
\item{NetM}{temp}
}
\value{
temp
}
\description{
<What does this function do>
}

