% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic_getEdgeProbMat.R
\name{getEdgeProbMat}
\alias{getEdgeProbMat}
\title{Extracts the edge probability matrix for a network model}
\usage{
getEdgeProbMat(NetM, mode = "prob")
}
\arguments{
\item{NetM}{[\code{\link{NetworkModel}}] :: Network Model}

\item{mode}{[char] :: Either "prob" or "group": if "prob", this function returns an edge probability matrix. Else, it returns a matrix where each cell gets a group id number}
}
\value{
[matrix] :: Edge probability matrix (or group index matrix)
}
\description{
Extracts the edge probability matrix for a network model
}

