% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_getNetType.R
\name{getNetType}
\alias{getNetType}
\title{Extract the type of network model}
\usage{
getNetType(Net)
}
\arguments{
\item{Net}{[\code{\link{NetworkModel}} OR \code{\link{NetworkStruct}}] :: Network object}
}
\value{
[char] :: type of network object
}
\description{
Allowable types are currently: "block", "tree", "latent", "random"
}

