% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic_getNnodes.R
\name{getNnodes}
\alias{getNnodes}
\title{Extracts the number of nodes in a network model}
\usage{
getNnodes(Net)
}
\arguments{
\item{Net}{NetworkModel or NetworkStruct object}
}
\value{
Numeric -- Number of nodes in network model
}
\description{
Extracts the number of nodes in a network model
}

