% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_edgegroupLik.R
\name{h_optim}
\alias{h_optim}
\title{<What does this function do>}
\usage{
h_optim(nparam, optim_tries, fn, gn, ...)
}
\arguments{
\item{nparam}{temp}

\item{optim_tries}{temp}

\item{fn}{temp}

\item{gn}{temp}

\item{...}{temp}
}
\value{
temp
}
\description{
<What does this function do>
}

