% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hCorr_paramToProb}
\alias{hCorr_paramToProb}
\title{Converts parameters to probabilities in correlation model}
\usage{
hCorr_paramToProb(rho, a, b = a)
}
\arguments{
\item{rho}{[double] :: correlation parameter}

\item{a}{[vector-double] :: network1 parameters}

\item{b}{[vector-double] :: network2 parameters}
}
\value{
[matrix-double] :: table with 4 columns -- probs of 11, 10, 01, 00
}
\description{
Converts parameters to probabilities in correlation model
}

