% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_edgegroupLik.R
\name{hCorr_probToParam}
\alias{hCorr_probToParam}
\title{Converts probabilities to parameters in correlation model}
\usage{
hCorr_probToParam(pt)
}
\arguments{
\item{pt}{[matrix-double] :: table with 4 columns -- probs of 11, 10, 01, 00}
}
\value{
[list] :: Named list of parameters in correlation model -- names are: 'a', 'b', 'lambda', 'rho'
}
\description{
Converts probabilities to parameters in correlation model
}

