% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_edgegroupLik.R
\name{hCorr_probToParam}
\alias{hCorr_probToParam}
\title{Converts probabilities to parameters in correlation model}
\usage{
hCorr_probToParam(pt)
}
\arguments{
\item{pt}{[] ::}
}
\value{
[] ::
}
\description{
Converts probabilities to parameters in correlation model
}

