% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hide_edges}
\alias{hide_edges}
\title{Hides a random set of edges}
\usage{
hide_edges(adjm, frac = 0.1, template = NULL, invert_template = FALSE)
}
\arguments{
\item{adjm}{[matrix-int] :: Input adjacency matrix}

\item{frac}{[double] :: Proportion of dyads to hide, rounded down to nearest number of dyads.}

\item{template}{[matrix] :: Matrix with NA's -- If this is provided, then the NA's in this matrix are copied into 'adjm'}

\item{invert_template}{[logical] :: If TRUE, the input template is inverted.}
}
\value{
[matrix-int] :: Adjacency matrix with hidden edges
}
\description{
Hides a random set of edges
}

