% Generated by roxygen2 (4.0.1): do not edit by hand
\name{HRG_closestAncestor}
\alias{HRG_closestAncestor}
\title{Compute ancestor_table from a tree}
\usage{
HRG_closestAncestor(tree)
}
\arguments{
\item{tree}{[\code{\link{NetworkModelHRG}}] :: Input tree}
}
\value{
[list] :: expanded_children & ancestor_table
}
\description{
Compute ancestor_table from a tree
}

