% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_compute_lrt.R
\name{is_prob}
\alias{is_prob}
\title{Checks vector for valid probabilities}
\usage{
is_prob(x)
}
\arguments{
\item{x}{[vector-double] :: Vector to check for probabilities}
}
\value{
[vector-logical] :: ith entry is TRUE if x[i] is a valid probability
}
\description{
Checks vector for valid probabilities
}

