% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_edgegroupLik.R
\name{llFx_calt}
\alias{llFx_calt}
\title{<What does this function do>}
\usage{
llFx_calt(t, C, n)
}
\arguments{
\item{t}{temp}

\item{C}{temp}

\item{n}{temp}
}
\value{
temp
}
\description{
<What does this function do>
}

