% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_edgegroupLik.R
\name{llFx_dendiff}
\alias{llFx_dendiff}
\title{Compute log-likelihood function for density-difference}
\usage{
llFx_dendiff(t, x, y, n)
}
\arguments{
\item{t}{[vector-double] :: c(b,a) - b is single value; a is vector}

\item{x}{[vector-int] :: edge counts in net 1}

\item{y}{[vector-int] :: edge counts in net 2}

\item{n}{[vector-int] :: total edge group size}
}
\value{
[double] :: value of log-likelihood
}
\description{
Compute log-likelihood function for density-difference
}

