% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_edgegroupLik.R
\name{llGrFx_dendiff}
\alias{llGrFx_dendiff}
\title{Compute gradient vector of log-likelihood in density-difference model}
\usage{
llGrFx_dendiff(t, x, y, n)
}
\arguments{
\item{t}{[vector-double] :: c(b,a) - b is single value; a is vector}

\item{x}{[vector-int] :: edge counts in net 1}

\item{y}{[vector-int] :: edge counts in net 2}

\item{n}{[vector-int] :: total edge group size}
}
\value{
[vector-double] :: gradient vector
}
\description{
Compute gradient vector of log-likelihood in density-difference model
}

