% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mult_bonferroni}
\alias{mult_bonferroni}
\title{Computes Bonferroni correction}
\usage{
mult_bonferroni(x)
}
\arguments{
\item{x}{[vector-double] :: P-values}
}
\value{
[double] :: Minimal p-value
}
\description{
Computes Bonferroni correction
}

