% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mult_quantileq}
\alias{mult_quantileq}
\title{Compute quantile test statistic (Q)}
\usage{
mult_quantileq(x, alpha = 0.05)
}
\arguments{
\item{x}{[vector-double] :: Input p-values}

\item{alpha}{[double] :: Alpha parameter in statistic}
}
\value{
[double] :: Test statistic
}
\description{
Compute quantile test statistic (Q)
}

