% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing_fx.R
\name{mult_quantiler}
\alias{mult_quantiler}
\title{Compute quantile test statistic (R)}
\usage{
mult_quantiler(x, alpha_min = 0.05, alpha_max = 1)
}
\arguments{
\item{x}{temp}

\item{alpha_min}{temp}

\item{alpha_max}{temp}
}
\value{
[double] :: Test statistic
}
\description{
Compute quantile test statistic (R)
}

