% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h_edgegroupLik.R
\name{reassign_edgegroup_prob}
\alias{reassign_edgegroup_prob}
\title{<What does this function do>}
\usage{
reassign_edgegroup_prob(NetM, ids, probs)
}
\arguments{
\item{NetM}{temp}

\item{ids}{temp}

\item{probs}{temp}
}
\value{
temp
}
\description{
<What does this function do>
}

