% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parameter_lists.R
\name{set_fit_param}
\alias{set_fit_param}
\title{Set model fitting parameters}
\usage{
set_fit_param(Ntries = 10, verbose = 0, SBM_start = "spectral",
  SBM_method = "mf", SBM_Nclass = 3, SBM_EM_Niter = 100,
  SBM_EM_stopthres = 1e-05)
}
\arguments{
\item{Ntries}{[int] :: Number of EM runs to optimize over}

\item{verbose}{[int] :: Verbosity level -- higher = more output}

\item{SBM_start}{[char] :: 'spectral' or 'random -- method to initialize starting state of EM algorithm}

\item{SBM_method}{[char] :: 'spectral' or 'mf' -- method to fit SBM: spectral does spectral clustering; mf does EM algo}

\item{SBM_Nclass}{[int] :: Number of classes in SBM}

\item{SBM_EM_Niter}{[int] :: Number of maximum iterations in EM}

\item{SBM_EM_stopthres}{[double] :: Threshold to declare convergence}
}
\value{
[list] :: A list of parameters
}
\description{
This function sets up the (list of) parameters to be passed into model fitting functions. There are default parameters that are used if this function is called with no arguments.
}

