% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_SBM.R
\name{specClust}
\alias{specClust}
\title{Runs spectral clustering on an input matrix}
\usage{
specClust(adjm, Nclass, Ntries)
}
\arguments{
\item{adjm}{[matrix-int] :: Input adjacency matrix}

\item{Nclass}{[int] :: Number of clusters}

\item{Ntries}{[int] :: Number of attempts}
}
\value{
[\code{\link{NetworkModel}}] :: Output bestfit model
}
\description{
Runs spectral clustering on an input matrix
}

