% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_SBM.R
\name{specClust}
\alias{specClust}
\title{Runs spectral clustering}
\usage{
specClust(evs, Nclass, NStart = 3)
}
\arguments{
\item{evs}{[matrix-numeric] :: Matrix of eigenvectors, with the columns as individual eigenvectors}

\item{Nclass}{[int] :: Number of classes to return for spectral clustering.}

\item{NStart}{[int] :: Number of random starts for the k-means algorithm}
}
\value{
[\code{\link{NetworkStruct}}] :: Output clustering's network structure
}
\description{
This requires input of eigenvectors. The eigenvectors can come from the Laplacian matrix, however.
}

