% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_SBM.R
\name{specClust}
\alias{specClust}
\title{Runs spectral clustering}
\usage{
specClust(adjm, evs, Nclass, Ntries, NStartPerTry = 2)
}
\arguments{
\item{adjm}{[matrix-int] :: Input adjacency matrix}

\item{evs}{[matrix-numeric] :: Matrix of eigenvectors, with the columns as individual eigenvectors}

\item{Nclass}{[vector-int] :: Number of classes to return for spectral clustering. If this is a vector, the best model is returned (with BIC penalization)}

\item{Ntries}{[int] :: Number of attempts (Number of times to run the k-means algorithm)}

\item{NStartPerTry}{[int] :: Number of random starts for the k-means algorithm}
}
\value{
[\code{\link{NetworkModel}}] :: Output bestfit model
}
\description{
This requires input of an adjacency matrix and eigenvectors. The eigenvectors can come from the Laplacian matrix, however. The input adjacency matrix can have missing data (it's only used to find the iteration that provides the highest likelihood)
}
\details{
This is also capable of searching over clusterings with different number of classes -- it picks the best model with respect to BIC.
}

