% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_SBM.R
\name{specClust}
\alias{specClust}
\title{<What does this function do>}
\usage{
specClust(adjm, Nclass)
}
\arguments{
\item{adjm}{temp}

\item{Nclass}{temp}
}
\value{
temp
}
\description{
<What does this function do>
}

